requirements.txt
gradio>=4.19.0
openai>=1.12.0
pypdf>=3.17.0
langchain>=0.1.0
langchain_community>=0.0.13
langchain_openai>=0.0.5
pyzotero>=1.5.5
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1 